refactor: Rename tests to allow adding a StaySeated test first
