+++ src/test/java/org/opentripplanner/raptor/moduletests/E03_NotAllowedConstrainedTransferTest.java
@@ -3,6 +3,7 @@ package org.opentripplanner.raptor.moduletests;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.raptor._data.transit.TestRoute.route;
 import static org.opentripplanner.raptor._data.transit.TestTripSchedule.schedule;
+import static org.opentripplanner.raptor.moduletests.support.RaptorModuleTestConfig.TC_STANDARD_REV;
 import static org.opentripplanner.raptor.moduletests.support.RaptorModuleTestConfig.multiCriteria;
 import static org.opentripplanner.raptor.moduletests.support.RaptorModuleTestConfig.standard;
 
@@ -19,17 +20,14 @@ import org.opentripplanner.raptor._data.transit.TestTripSchedule;
 import org.opentripplanner.raptor.api.request.RaptorRequestBuilder;
 import org.opentripplanner.raptor.configure.RaptorConfig;
 import org.opentripplanner.raptor.moduletests.support.RaptorModuleTestCase;
-import org.opentripplanner.raptor.spi.DefaultSlackProvider;
 
 /**
  * FEATURE UNDER TEST
  * <p>
- * Raptor should return a path if it exists when a transfer is only possible because it is
- * guaranteed/stay-seated. A guarantied transfer should be able even if there is zero time to do the
- * transfer. In these cases the transfer-slack should be ignored and the connection should be
- * possible.
+ * Raptor should NOT return a path with a NOT-ALLOWED transfer, instead it should try to find
+ * another option.
  */
-public class E01_GuaranteedTransferTest implements RaptorTestConstants {
+public class E03_NotAllowedConstrainedTransferTest implements RaptorTestConstants {
 
   private final TestTransitData data = new TestTransitData();
   private final RaptorRequestBuilder<TestTripSchedule> requestBuilder = new RaptorRequestBuilder<>();
@@ -45,16 +43,29 @@ public class E01_GuaranteedTransferTest implements RaptorTestConstants {
   @BeforeEach
   public void setup() {
     var r1 = route("R1", STOP_A, STOP_B).withTimetable(schedule("0:02 0:05"));
-    var r2 = route("R2", STOP_B, STOP_C).withTimetable(schedule("0:05 0:10"));
+    var r2 = route("R2", STOP_B, STOP_C)
+      .withTimetable(
+        schedule("0:10 0:15"),
+        // Add another schedule - should not be used even if the not-allowed is
+        // attached to the first one - not-allowed in Raptor apply to the Route.
+        // The trip/timetable search should handle not-allowed on trip level.
+        schedule("0:12 0:17")
+      );
+    var r3 = route("R3", STOP_B, STOP_C).withTimetable(schedule("0:15 0:20"));
 
-    var tripA = r1.timetable().getTripSchedule(0);
-    var tripB = r2.timetable().getTripSchedule(0);
+    var tripR1a = r1.timetable().getTripSchedule(0);
+    var tripR2a = r2.timetable().getTripSchedule(0);
 
-    data.withRoutes(r1, r2);
-    data.withGuaranteedTransfer(tripA, STOP_B, tripB, STOP_B);
+    data.withRoutes(r1, r2, r3);
+
+    // Apply not-allowed on the first trip from R1 and R2 - when found this will apply to
+    // the second trip in R2 as well. This is of cause not a correct way to implement the
+    // transit model, a not-allowed transfer should apply to ALL trips if constraint is passed
+    // to raptor.
+    data.withConstrainedTransfer(tripR1a, STOP_B, tripR2a, STOP_B, TestTransitData.TX_NOT_ALLOWED);
     data.mcCostParamsBuilder().transferCost(100);
 
-    // NOTE! No search-window is set.
+    // NOTE! No search-window set
     requestBuilder
       .searchParams()
       .constrainedTransfers(true)
@@ -64,21 +75,17 @@ public class E01_GuaranteedTransferTest implements RaptorTestConstants {
       .latestArrivalTime(T00_30)
       .timetable(true);
 
-    // Make sure the slack have values which prevent the normal from happening transfer.
-    // The test scenario have zero seconds to transfer, so any slack will do.
-    data.withSlackProvider(new DefaultSlackProvider(D30s, D20s, D10s));
-
     ModuleTestDebugLogging.setupDebugLogging(data, requestBuilder);
   }
 
   static List<RaptorModuleTestCase> testCases() {
     var path =
-      "Walk 30s ~ A ~ BUS R1 0:02 0:05 ~ B ~ BUS R2 0:05 0:10 ~ C ~ Walk 30s " +
-      "[0:01:10 0:10:40 9m30s 1tx $1230]";
+      "Walk 30s ~ A ~ BUS R1 0:02 0:05 ~ B ~ BUS R3 0:15 0:20 ~ C ~ Walk 30s " +
+      "[0:01:30 0:20:30 19m 1tx $2500]";
     return RaptorModuleTestCase
       .of()
-      .addMinDuration("9m30s", TX_1, T00_00, T00_30)
-      .add(standard(), PathUtils.withoutCost(path))
+      .addMinDuration("9m", TX_1, T00_00, T00_30)
+      .add(standard().not(TC_STANDARD_REV), PathUtils.withoutCost(path))
       .add(multiCriteria(), path)
       .build();
   }
